+++ lib/cli/src/commands/run.rs
@@ -1,4 +1,4 @@
-use crate::cli::{SplitVersion, SplitVersionError};
+use crate::split_version::{SplitVersion, SplitVersionError};
 use crate::common::get_cache_dir;
 #[cfg(feature = "debug")]
 use crate::logging;
@@ -16,7 +16,7 @@ use wasmer::FunctionEnv;
 use wasmer::*;
 #[cfg(feature = "cache")]
 use wasmer_cache::{Cache, FileSystemCache, Hash};
-use wasmer_registry::PackageDownloadInfo;
+use wasmer_registry::{PackageDownloadInfo, SplitVersion};
 use wasmer_types::Type as ValueType;
 #[cfg(feature = "webc_runner")]
 use wasmer_wasi::runners::{Runner, WapmContainer};
