+++ lib/registry/Cargo.toml
@@ -15,6 +15,7 @@ serde = { version = "1.0.145", features = ["derive"] }
 anyhow = "1.0.65" 
 reqwest = { version = "0.11.12", default-features = false, features = ["rustls-tls", "blocking", "multipart", "json", "stream"] }
 futures-util = "0.3.25"
+regex = "1.7.0"
 whoami = "1.2.3" 
 serde_json = "1.0.85"
 url = "2.3.1"
